fixed horrible bug

+++ a4.cl
@@ -1,4 +1,4 @@
-__kernel void oclgrind(__local char* grid, __local float* gridSize) {
+__kernel void oclgrind(__local char* grid, __local int* gridSize) {
   
    
 }
